refactor: split terminal side chat workflow
