+++ apps/web/src/features/terminal/lib/terminal-side-chat.ts
@@ -1,0 +1,198 @@
+import type {
+  TerminalSideChatRecord,
+  TerminalSideChatStatus,
+  TerminalSideContextCaptureResponse,
+} from "@/api/ws-api";
+import type { TerminalPaneAgent } from "@/features/terminal/types";
+
+export type SourceSurfaceKind = "terminal_pane" | "canvas_terminal";
+
+export type LocalSideChatRecord = TerminalSideChatRecord & {
+  agent?: TerminalPaneAgent;
+  hasSentInitialCommand?: boolean;
+  initialCommand?: string;
+  isNew: boolean;
+  sessionId: string;
+};
+
+type LegacySideChatStatus = TerminalSideChatStatus | "visible" | "closed" | string;
+
+const BRIGHT_SIDE_CHAT_COLORS = [
+  "#06b6d4",
+  "#22c55e",
+  "#f59e0b",
+  "#ef4444",
+  "#8b5cf6",
+  "#ec4899",
+  "#14b8a6",
+  "#84cc16",
+  "#3b82f6",
+  "#f97316",
+];
+
+export function sideChatTabLabel(
+  record: LocalSideChatRecord,
+  index: number,
+  fallbackTitle: string,
+): string {
+  return record.agent?.label?.trim() || `${fallbackTitle} ${index + 1}`;
+}
+
+export function sideChatRecordMatchesSource(
+  record: TerminalSideChatRecord,
+  sourceSurfaceRefJson: string | null,
+  sourceTmuxWindowName?: string | null,
+): boolean {
+  if (sourceSurfaceRefJson && record.source_surface_ref_json) {
+    return record.source_surface_ref_json === sourceSurfaceRefJson;
+  }
+  return record.source_tmux_window_name === sourceTmuxWindowName;
+}
+
+export function mergeSideChatRecords(
+  current: LocalSideChatRecord[],
+  incoming: LocalSideChatRecord[],
+) {
+  const currentById = new Map(current.map((record) => [record.side_chat_id, record]));
+  const incomingIds = new Set(incoming.map((record) => record.side_chat_id));
+  const merged = incoming.map((record) => ({
+    ...record,
+    ...currentById.get(record.side_chat_id),
+    ...record,
+  }));
+  for (const record of current) {
+    if (!incomingIds.has(record.side_chat_id) && record.isNew) {
+      merged.push(record);
+    }
+  }
+  return merged;
+}
+
+export function toSideChatDto(record: LocalSideChatRecord): TerminalSideChatRecord {
+  return {
+    side_chat_id: record.side_chat_id,
+    workspace_id: record.workspace_id,
+    project_name: record.project_name,
+    workspace_name: record.workspace_name,
+    source_pane_id: record.source_pane_id,
+    source_tmux_window_name: record.source_tmux_window_name,
+    source_surface_kind: record.source_surface_kind,
+    source_surface_ref_json: record.source_surface_ref_json,
+    side_tmux_window_name: record.side_tmux_window_name,
+    agent_ref_json: record.agent_ref_json,
+    color_hex: record.color_hex,
+    status: normalizeSideChatStatus(record.status),
+    created_at: record.created_at,
+    updated_at: record.updated_at,
+  };
+}
+
+export function normalizeSideChatStatus(status: LegacySideChatStatus): TerminalSideChatStatus {
+  if (status === "visible") return "open";
+  if (status === "closed") return "closing";
+  if (status === "open" || status === "hidden" || status === "closing") return status;
+  return "hidden";
+}
+
+export function isSideChatOpen(status: LegacySideChatStatus): boolean {
+  return normalizeSideChatStatus(status) === "open";
+}
+
+export function isSideChatClosing(status: LegacySideChatStatus): boolean {
+  return normalizeSideChatStatus(status) === "closing";
+}
+
+export function pickUniqueBrightColor(existingColors: string[]) {
+  const used = new Set(existingColors.map((color) => color.toLowerCase()));
+  const available = BRIGHT_SIDE_CHAT_COLORS.filter((color) => !used.has(color.toLowerCase()));
+  if (available.length > 0) {
+    return available[Math.floor(Math.random() * available.length)];
+  }
+  for (let attempt = 0; attempt < 20; attempt += 1) {
+    const hue = Math.floor(Math.random() * 360);
+    const color = hslToHex(hue, 78, 54);
+    if (!used.has(color.toLowerCase())) return color;
+  }
+  return hslToHex(Math.floor(Math.random() * 360), 78, 54);
+}
+
+export function buildSideChatPrompt({
+  capture,
+  sourceTmuxWindowName,
+  userPrompt,
+}: {
+  capture: TerminalSideContextCaptureResponse;
+  sourceTmuxWindowName: string;
+  userPrompt: string;
+}) {
+  const metadata = [
+    `Source terminal: ${sourceTmuxWindowName}`,
+    `Captured lines: ${capture.captured_lines}`,
+    `Captured bytes: ${capture.captured_bytes}/${capture.prompt_budget_bytes}`,
+  ];
+  if (capture.omitted_older_bytes > 0 || capture.omitted_middle_bytes > 0 || capture.truncated_bytes) {
+    metadata.push("Capture was bounded; omitted content may exist outside this excerpt.");
+  }
+
+  return [
+    "You are continuing in a side chat forked from an Atmos terminal.",
+    "Use the captured terminal context below as background. Do not assume it is complete.",
+    "",
+    metadata.join("\n"),
+    "",
+    "Captured terminal context:",
+    "```text",
+    capture.text,
+    "```",
+    "",
+    "User prompt:",
+    userPrompt.trim(),
+  ].join("\n");
+}
+
+export function parseAgentRef(value: string | null | undefined): TerminalPaneAgent | undefined {
+  if (!value) return undefined;
+  try {
+    const parsed = JSON.parse(value) as Partial<TerminalPaneAgent>;
+    if (
+      typeof parsed.id === "string" &&
+      typeof parsed.label === "string" &&
+      typeof parsed.command === "string" &&
+      (parsed.iconType === "built-in" || parsed.iconType === "custom")
+    ) {
+      return {
+        id: parsed.id,
+        label: parsed.label,
+        command: parsed.command,
+        iconType: parsed.iconType,
+        pipeCommand: typeof parsed.pipeCommand === "string" ? parsed.pipeCommand : undefined,
+      };
+    }
+  } catch {
+    return undefined;
+  }
+  return undefined;
+}
+
+function hslToHex(h: number, s: number, l: number) {
+  const saturation = s / 100;
+  const lightness = l / 100;
+  const c = (1 - Math.abs(2 * lightness - 1)) * saturation;
+  const x = c * (1 - Math.abs(((h / 60) % 2) - 1));
+  const m = lightness - c / 2;
+  const [r, g, b] =
+    h < 60
+      ? [c, x, 0]
+      : h < 120
+        ? [x, c, 0]
+        : h < 180
+          ? [0, c, x]
+          : h < 240
+            ? [0, x, c]
+            : h < 300
+              ? [x, 0, c]
+              : [c, 0, x];
+  return `#${[r, g, b]
+    .map((value) => Math.round((value + m) * 255).toString(16).padStart(2, "0"))
+    .join("")}`;
+}
